* Se han cambiado el timer TCNT0 (8bits) por el TCNT1 (16bits).
* Eliminación de AHRS.{cpp,h} y declarations.h.
* Adaptado Makefile para el uso de frame y la eliminacion de declarations.
* Desactiva interrupción por overflow del timer16b.

Conflicts:
	algorithm.cpp
	algorithm.h
	main.cpp
	quaternion.h
	vector.cpp
	vector.h

+++ gyroscope.h
@@ -25,8 +25,8 @@ class Gyroscope{
 
 		bool getData(Vector<float> &vector, float &temperature);
 
-		uint8_t deltaT;
-		uint8_t deltaT_old;
+		uint16_t deltaT;
+		uint16_t deltaT_old;
 
 	private:
 		bool dataReady();
